backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
twilio==9.0.4
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.3
